TrendCommand: Ensure map is sorted by year
TrendCommand uses a generic Map to store the year values. As such, when
the map is used to generate the JSON file later, this results in the
year-count-pair to be unordered.

This implies that we cannot be sure about the actual results that will
be generated in our tests.

Other considerations:
I have thought of updating MapUtility’s methods to return a LinkedHashMap
instead. However, this does not work as LinkedHashMap only maintains
insertion order and there is no guarantee in our code that we will insert
each year in ascending year order.